staging: etnaviv: separate GPU pipes from execution state
While this isn't the case on i.MX6 a single GPU pipe can have
multiple rendering backend states, which can be selected by the
pipe switch command, so there is no strict mapping between the
user "pipes" and the PIPE_2D/PIPE_3D execution states.

We need to respect this in the public userspace API. The exported
pipes are the GC cores which may be able to support one or more
execution states. The information which execution states are supported
on a given pipe is available to userspace through the features0 param.

Userspace is responsible to choose a matching pipe when deciding which
pipe to use for a specific task.

The submit ioctl now takes one more parameter for userspace to specify
which execution state it expects a pipe to be in when starting execution
of the command buffers. This allows the kernel to insert pipe switch
commands only when really needed while maintaining separation between
processes by making sure that no process can leave the pipe behind in
an unexpected state, potentially hanging the next one to use the GPU.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>