staging: etnaviv: align command stream size to 64 bit
It is legal for the userspace to pass in a command stream of a size
aligned to 32 bit, if that is where the last user command ends. The
kernel then needs to insert a LINK command at the end of the stream,
which needs to be aligned to 64 bit, so the kernel may insert an
additional 32bits of padding in the stream. Align the stream size
to account for that in the size and command stream validator checks.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>