staging: etnaviv: add support for offset physical memory
On iMX6, memory starts at 256MB physical, and if we have 2GB of memory,
this causes it to extend beyond 0x80000000.  All memory is available for
DMA coherent allocations, which can result in the command buffers being
allocated from addresses above 0x80000000.

However, the Vivante GPU requires that command buffers are in the linear
space (first 2GB of GPU addresses.)  Trying to program an address with
bit 31 set results in it being ignored.

The Vivante GPU has a solution to this: they provide a set of memory
base address registers, which are added to the GPU linear space address
before appearing on the bus.  If we program these to the base address of
physical memory, we can use all 2GB of RAM.

There are other reasons to do this: firstly, other SoCs may start their
physical memory at other bus addresses, which may further reduce the
range of physical addresses for command buffers.  Secondly, it prevents
the GPU being able to access addresses below RAM, which in the iMX6
case are hardware registers.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>